[No Debug] Introduced standardized types; Some refactoring of planner. TODO: Replace TABS for spaces

+++ include/planner/genTrajectory.h
@@ -61,7 +61,7 @@ private:
 	bool m_isControlStarted;
 	bool m_isFirstTimePass;
 
-	double m_startTime;
+	float32 m_startTime;
 
 	TrajectoryParameters m_parameters;
 };
